requirements.txt
anyio==4.4.0
APScheduler==3.10.4
certifi==2024.6.2
charset-normalizer==3.3.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
python-telegram-bot==21.3
pytz==2024.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tzlocal==5.2
urllib3==2.2.2
